+++ apps/explorer/src/components/SearchButton.tsx
@@ -2,17 +2,20 @@ import { SearchIcon } from "@/icons/SearchIcon";
 import { StyledWrapper } from "./TxHashInput";
 import { styled, useTheme } from "@/styled-components";
 import { Text } from "@/components/Typography";
+import { loadingPulseAnimation } from "@/components/Container";
 
 export const SearchButton = ({
   size = "normal",
   onClick,
   iconOnly = false,
   isSearchable = false,
+  isLoading = false,
 }: {
   size?: "normal" | "small";
   onClick?: () => void;
   iconOnly?: boolean;
   isSearchable?: boolean;
+  isLoading?: boolean;
 }) => {
   const theme = useTheme();
   return (
@@ -22,6 +25,7 @@ export const SearchButton = ({
       onClick={onClick}
       iconOnly={iconOnly}
       isSearchable={isSearchable}
+      isLoading={isLoading}
     >
       <StyledText size={size} iconOnly={iconOnly} isSearchable={isSearchable}>
         Search transaction
@@ -34,12 +38,16 @@ export const SearchButton = ({
 const SearchWrapper = styled(StyledWrapper)<{
   iconOnly?: boolean;
   isSearchable?: boolean;
+  isLoading?: boolean;
 }>`
   background-color: ${(props) => props.isSearchable ? props.theme.brandColor : props.theme.primary.background.normal};
   width: ${(props) => (props.size === "normal" ? "64px" : "48px")};
   height: ${(props) => (props.size === "normal" ? "64px" : "48px")};
   justify-content: center;
   flex: none;
+  ${({ isLoading }) => isLoading && loadingPulseAnimation({
+    active: true,
+  })}
   @media (max-width: 1300px) {
     width: ${(props) =>
       props.iconOnly ? (props.size === "normal" ? "64px" : "48px") : "auto"};