+++ packages/widget/CHANGELOG.md
@@ -1,5 +1,33 @@
 ## 0.0.15
 
+## 3.14.20
+
+### Patch Changes
+
+- 12f2603: add explorer loading state
+
+## 3.14.19
+
+### Patch Changes
+
+- 6aafa55: fix webcomponent
+
+## 3.14.18
+
+### Patch Changes
+
+- ae6abc9: update deps and support cosmos mevm
+- Updated dependencies [ae6abc9]
+  - @skip-go/client@1.5.11
+
+## 3.14.17
+
+### Patch Changes
+
+- 69fda94: override asset image
+- Updated dependencies [69fda94]
+  - @skip-go/client@1.5.10
+
 ## 3.14.16
 
 ### Patch Changes